Crate starchart

A simple database system that allows the use of multiple different backends.

Modules

action

The action structs for CRUD operations.

backend

The backend that fetches and provides data for the Starchart.

error

The different errors within the crate.

Structs

Action

An Action for easy CRUD operations within a Starchart.

Error

An error that occurred within the crate.

Starchart

The base structure for managing data.

Traits

Entry

A marker trait for use within the Starchart.

IndexEntry

An indexable entry, used for any Entry that can be indexed by a Key that it owns.

Key

The key trait to be implemented on Entry to allow an easy way to get keys.

Type Definitions

Result

A type alias for a Result that wraps around Error.

Derive Macros

IndexEntryderive

The helper derive macro for easily implementing IndexEntry.