Module starchart::action

The action structs for CRUD operations.

Structs

Action

An Action for easy CRUD operations within a Starchart.

CreateOperation

Marker type for a Create operation.

DeleteOperation

Marker type for a Delete operation.

EntryTarget

Marker type for an entry operation.

ReadOperation

Marker type for a Read operation.

TableTarget

Marker type for a table operation.

UpdateOperation

Marker type for an Update operation.

Enums

ActionKind

The type of CRUD action to perform

OperationTarget

The target of the CRUD operation.

Traits

ActionRunner

The marker trait for all action runs, this trait should not be used and is only used to make the return type of Starchart::run easily known.

CrudOperation

A sealed marker trait for helping an Action represent what type of operation will occur.

OpTarget

A sealed marker trait for helping an Action represent what type of target the operation will cover.

Type Definitions

CreateEntryAction

A type alias for an Action with CreateOperation and EntryTarget as the parameters.

CreateTableAction

A type alias for an Action with CreateOperation and TableTarget as the parameters.

DeleteEntryAction

A type alias for an Action with DeleteOperation and EntryTarget as the parameters.

DeleteTableAction

A type alias for an Action with DeleteOperation and TableTarget as the parameters.

ReadEntryAction

A type alias for an Action with ReadOperation and EntryTarget as the parameters.

ReadTableAction

A type alias for an Action with ReadOperation and TableTarget as the parameters.

UpdateEntryAction

A type alias for an Action with UpdateOperation and EntryTarget as the parameters.

UpdateTableAction

A type alias for an Action with UpdateOperation and TableTarget as the parameters.