Enum starchart::action::ActionResult

pub enum ActionResult<R> {
    Create,
    SingleRead(Option<R>),
    MultiRead(Vec<R>),
    Update,
    Delete(bool),
}

A custom Result type that allows the Action::run method to return multiple different types.

Variants

Create

Indicates a CreateOperation was performed.

SingleRead(Option<R>)

Tuple Fields

0: Option<R>

Indicates a ReadOperation was performed on an EntryTarget.

This value can be retrieved with Self::unwrap_single_read.

MultiRead(Vec<R>)

Tuple Fields

0: Vec<R>

Indicates a ReadOperation was performed on a TableTarget.

This value can be retrieved with Self::unwrap_multi_read.

Update

Indicates an UpdateOperation was performed.

Delete(bool)

Tuple Fields

0: bool

Indicates a DeleteOperation was performed.

The inner value can be retrieved with Self::unwrap_delete to determine whether the delete was successful.

Implementations

impl<R> ActionResult<R>

pub const fn is_create(&self) -> bool

Returns true if the result is Create.

Examples found in repository

src/action/result.rs (line 103)

≺ ≻↕

101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124

pub fn unwrap_create(self) {
		assert!(
			self.is_create(),
			"called `ActionResult::create` on a `{}` value",
			self
		);
	}

	/// Unwraps the [`ActionResult::Create`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`Create`] will cause `UB`.
	///
	/// [`Create`]: Self::Create
	#[inline]
	#[track_caller]
	pub unsafe fn unwrap_create_unchecked(self) {
		debug_assert!(self.is_create());
		if let Self::Create = self {
		} else {
			unreachable_unchecked()
		}
	}

pub const fn is_single_read(&self) -> bool

Returns true if the result is SingleRead.

Examples found in repository

src/action/result.rs (line 73)

≺ ≻↕

72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159

pub const fn is_read(&self) -> bool {
		self.is_single_read() || self.is_multi_read()
	}

	/// Returns [`true`] if the result is an [`Update`].
	///
	/// [`Update`]: Self::Update
	pub const fn is_update(&self) -> bool {
		matches!(self, Self::Update)
	}

	/// Returns [`true`] if the result is a [`Delete`].
	///
	/// [`Delete`]: Self::Delete
	pub const fn is_delete(&self) -> bool {
		matches!(self, Self::Delete(_))
	}
}

impl<R: Entry> ActionResult<R> {
	/// Unwraps the [`ActionResult`] and asserts it's a create
	///
	/// # Panics
	///
	/// This method panics if [`Self`] is anything other than [`Create`].
	///
	/// [`Create`]: Self::Create
	#[track_caller]
	#[inline]
	pub fn unwrap_create(self) {
		assert!(
			self.is_create(),
			"called `ActionResult::create` on a `{}` value",
			self
		);
	}

	/// Unwraps the [`ActionResult::Create`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`Create`] will cause `UB`.
	///
	/// [`Create`]: Self::Create
	#[inline]
	#[track_caller]
	pub unsafe fn unwrap_create_unchecked(self) {
		debug_assert!(self.is_create());
		if let Self::Create = self {
		} else {
			unreachable_unchecked()
		}
	}

	/// Unwraps the [`ActionResult`] and asserts it's a single read.
	///
	/// # Panics
	///
	/// This method panics if [`Self`] is anything other than [`SingleRead`].
	///
	/// [`SingleRead`]: Self::SingleRead
	#[track_caller]
	#[inline]
	pub fn unwrap_single_read(self) -> Option<R> {
		if let Self::SingleRead(v) = self {
			v
		} else {
			panic!("called `ActionResult::single_read` on a `{}` value", self);
		}
	}

	/// Unwraps the [`ActionResult::SingleRead`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`SingleRead`] will cause `UB`.
	///
	/// [`SingleRead`]: Self::SingleRead
	#[track_caller]
	#[inline]
	pub unsafe fn unwrap_single_read_unchecked(self) -> Option<R> {
		debug_assert!(self.is_single_read());
		if let Self::SingleRead(v) = self {
			v
		} else {
			unreachable_unchecked()
		}
	}

pub const fn is_multi_read(&self) -> bool

Returns true if the result is MultiRead.

Examples found in repository

src/action/result.rs (line 73)

≺ ≻↕

 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192

pub const fn is_read(&self) -> bool {
		self.is_single_read() || self.is_multi_read()
	}

	/// Returns [`true`] if the result is an [`Update`].
	///
	/// [`Update`]: Self::Update
	pub const fn is_update(&self) -> bool {
		matches!(self, Self::Update)
	}

	/// Returns [`true`] if the result is a [`Delete`].
	///
	/// [`Delete`]: Self::Delete
	pub const fn is_delete(&self) -> bool {
		matches!(self, Self::Delete(_))
	}
}

impl<R: Entry> ActionResult<R> {
	/// Unwraps the [`ActionResult`] and asserts it's a create
	///
	/// # Panics
	///
	/// This method panics if [`Self`] is anything other than [`Create`].
	///
	/// [`Create`]: Self::Create
	#[track_caller]
	#[inline]
	pub fn unwrap_create(self) {
		assert!(
			self.is_create(),
			"called `ActionResult::create` on a `{}` value",
			self
		);
	}

	/// Unwraps the [`ActionResult::Create`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`Create`] will cause `UB`.
	///
	/// [`Create`]: Self::Create
	#[inline]
	#[track_caller]
	pub unsafe fn unwrap_create_unchecked(self) {
		debug_assert!(self.is_create());
		if let Self::Create = self {
		} else {
			unreachable_unchecked()
		}
	}

	/// Unwraps the [`ActionResult`] and asserts it's a single read.
	///
	/// # Panics
	///
	/// This method panics if [`Self`] is anything other than [`SingleRead`].
	///
	/// [`SingleRead`]: Self::SingleRead
	#[track_caller]
	#[inline]
	pub fn unwrap_single_read(self) -> Option<R> {
		if let Self::SingleRead(v) = self {
			v
		} else {
			panic!("called `ActionResult::single_read` on a `{}` value", self);
		}
	}

	/// Unwraps the [`ActionResult::SingleRead`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`SingleRead`] will cause `UB`.
	///
	/// [`SingleRead`]: Self::SingleRead
	#[track_caller]
	#[inline]
	pub unsafe fn unwrap_single_read_unchecked(self) -> Option<R> {
		debug_assert!(self.is_single_read());
		if let Self::SingleRead(v) = self {
			v
		} else {
			unreachable_unchecked()
		}
	}

	/// Unwraps the [`ActionResult`] and asserts it's a multi-read.
	///
	/// # Panics
	///
	/// This method panics if [`Self`] is anything other than [`MultiRead`].
	///
	/// [`MultiRead`]: Self::MultiRead
	#[track_caller]
	#[inline]
	pub fn unwrap_multi_read<I: FromIterator<R>>(self) -> I {
		if let Self::MultiRead(v) = self {
			v.into_iter().collect()
		} else {
			panic!("called `ActionResult::multi_read` on a `{}` value", self)
		}
	}

	/// Unwraps the [`ActionResult::MultiRead`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than a [`MultiRead`] will cause `UB`.
	///
	/// [`MultiRead`]: Self::MultiRead
	pub unsafe fn unwrap_multi_read_unchecked<I: FromIterator<R>>(self) -> I {
		debug_assert!(self.is_multi_read());
		if let Self::MultiRead(v) = self {
			v.into_iter().collect()
		} else {
			unreachable_unchecked()
		}
	}

pub const fn is_read(&self) -> bool

Returns true if the result is either a SingleRead or a MultiRead.

pub const fn is_update(&self) -> bool

Returns true if the result is an Update.

Examples found in repository

src/action/result.rs (line 205)

≺ ≻↕

203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224

pub fn unwrap_update(self) {
		assert!(
			self.is_update(),
			"called `ActionResult::update` on a `{}` value",
			self
		);
	}

	/// Unwraps the [`ActionResult::Update`] value, without checking if the value is any other type.
	///
	/// # Safety
	///
	/// Calling this method on anything other than an [`Update`] will cause `UB`.
	///
	/// [`Update`]: Self::Update
	pub unsafe fn unwrap_update_unchecked(self) {
		debug_assert!(self.is_update());
		if let Self::Update = self {
		} else {
			unreachable_unchecked()
		}
	}

pub const fn is_delete(&self) -> bool

Returns true if the result is a Delete.

Examples found in repository

src/action/result.rs (line 251)

250
251
252
253
254
255
256
257

pub unsafe fn unwrap_delete_unchecked(self) -> bool {
		debug_assert!(self.is_delete());
		if let Self::Delete(b) = self {
			b
		} else {
			unreachable_unchecked()
		}
	}

impl<R: Entry> ActionResult<R>

pub fn unwrap_create(self)

Unwraps the ActionResult and asserts it’s a create

Panics

This method panics if Self is anything other than Create.

Examples found in repository

examples/get_all.rs (line 51)

≺ ≻↕

43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101

fn main() -> Result<()> {
	let chart = Starchart::new(JsonBackend::new("./examples/get_all_db")?).await?;

	{
		let mut action: CreateTableAction<Settings> = Action::new();
		action.set_table("foo".to_owned());

		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	let mut read_all_action: ReadTableAction<Settings> = Action::new();
	read_all_action.set_table("foo".to_owned());

	// let mut values: Vec<Settings> = chart.run(read_all_action).await??;
	let mut values = read_all_action
		.run(&chart)
		.await?
		.unwrap_multi_read::<Vec<_>>();

	values.sort();

	assert_eq!(
		values,
		vec![
			Settings {
				id: 1,
				name: "John Doe".to_owned(),
				age: 21
			},
			Settings {
				id: 2,
				name: "Ferris".to_owned(),
				age: 42
			},
			Settings {
				id: 3,
				name: "The Queen".to_owned(),
				age: 73
			}
		]
	);

	Ok(())
}

examples/basic.rs (line 40)

≺ ≻↕

30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80

fn main() -> Result<()> {
	// Create an initialize the database.
	let chart = Starchart::new(MemoryBackend::default()).await?;

	// Create and run an action to create the "foo" table with the Settings struct.
	let mut create_table_action: CreateTableAction<Settings> = Action::new();

	create_table_action.set_table("foo".to_owned());

	// chart.run(create_table_action).await??;
	create_table_action.run(&chart).await?.unwrap_create();

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Get a single entry.

	let the_queen = {
		// Action type helpers are named after their CRUD counterparts; Create, Read, Update, and Delete.
		let mut action: ReadEntryAction<Settings> = Action::new();
		action.set_key(&3_u64).set_table("foo".to_owned());

		action
			.run(&chart)
			.await?
			.unwrap_single_read()
			.expect("the queen has fallen!")
	};

	assert_eq!(
		the_queen,
		Settings {
			id: 3,
			name: "The Queen".to_owned(),
			age: 73
		}
	);

	Ok(())
}

pub unsafe fn unwrap_create_unchecked(self)

Unwraps the ActionResult::Create value, without checking if the value is any other type.

Safety

Calling this method on anything other than a Create will cause UB.

pub fn unwrap_single_read(self) -> Option<R>

Unwraps the ActionResult and asserts it’s a single read.

Panics

This method panics if Self is anything other than SingleRead.

Examples found in repository

examples/basic.rs (line 66)

↕

30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80

fn main() -> Result<()> {
	// Create an initialize the database.
	let chart = Starchart::new(MemoryBackend::default()).await?;

	// Create and run an action to create the "foo" table with the Settings struct.
	let mut create_table_action: CreateTableAction<Settings> = Action::new();

	create_table_action.set_table("foo".to_owned());

	// chart.run(create_table_action).await??;
	create_table_action.run(&chart).await?.unwrap_create();

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Get a single entry.

	let the_queen = {
		// Action type helpers are named after their CRUD counterparts; Create, Read, Update, and Delete.
		let mut action: ReadEntryAction<Settings> = Action::new();
		action.set_key(&3_u64).set_table("foo".to_owned());

		action
			.run(&chart)
			.await?
			.unwrap_single_read()
			.expect("the queen has fallen!")
	};

	assert_eq!(
		the_queen,
		Settings {
			id: 3,
			name: "The Queen".to_owned(),
			age: 73
		}
	);

	Ok(())
}

pub unsafe fn unwrap_single_read_unchecked(self) -> Option<R>

Unwraps the ActionResult::SingleRead value, without checking if the value is any other type.

Safety

Calling this method on anything other than a SingleRead will cause UB.

pub fn unwrap_multi_read<I: FromIterator<R>>(self) -> I

Unwraps the ActionResult and asserts it’s a multi-read.

Panics

This method panics if Self is anything other than MultiRead.

Examples found in repository

examples/get_all.rs (line 75)

↕

43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101

fn main() -> Result<()> {
	let chart = Starchart::new(JsonBackend::new("./examples/get_all_db")?).await?;

	{
		let mut action: CreateTableAction<Settings> = Action::new();
		action.set_table("foo".to_owned());

		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	let mut read_all_action: ReadTableAction<Settings> = Action::new();
	read_all_action.set_table("foo".to_owned());

	// let mut values: Vec<Settings> = chart.run(read_all_action).await??;
	let mut values = read_all_action
		.run(&chart)
		.await?
		.unwrap_multi_read::<Vec<_>>();

	values.sort();

	assert_eq!(
		values,
		vec![
			Settings {
				id: 1,
				name: "John Doe".to_owned(),
				age: 21
			},
			Settings {
				id: 2,
				name: "Ferris".to_owned(),
				age: 42
			},
			Settings {
				id: 3,
				name: "The Queen".to_owned(),
				age: 73
			}
		]
	);

	Ok(())
}

pub unsafe fn unwrap_multi_read_unchecked<I: FromIterator<R>>(self) -> I

Unwraps the ActionResult::MultiRead value, without checking if the value is any other type.

Safety

Calling this method on anything other than a MultiRead will cause UB.

pub fn unwrap_update(self)

Unwraps the ActionResult and asserts it’s an update.

Panics

This method panics if Self is anything other than Update.

pub unsafe fn unwrap_update_unchecked(self)

Unwraps the ActionResult::Update value, without checking if the value is any other type.

Safety

Calling this method on anything other than an Update will cause UB.

pub fn unwrap_delete(self) -> bool

Unwraps the ActionResult and asserts it’s a delete.

Panics

This method panics if Self is anything other than Delete.

pub unsafe fn unwrap_delete_unchecked(self) -> bool

Unwraps the ActionResult::Delete value, without checking if the value is any other type.

Safety

Calling this method on anything other than a Delete will cause UB.