Struct starchart::action::Action

pub struct Action<S, C, T> { /* fields omitted */ }

An Action for easy CRUD operations within a Starchart.

Implementations

impl<S, C, T> Action<S, C, T>

pub const fn new() -> Self

Creates a new Action with the specified operation.

pub fn table(&self) -> Option<&str>

Get a reference to the currently set table.

pub fn key(&self) -> Option<&str>

Get a reference to the currently set key.

impl<S: Entry, C: CrudOperation, T: OperationTarget> Action<S, C, T>

pub fn data(&self) -> Option<&S>

Get a reference to the currently set data.

pub fn kind(&self) -> ActionKind

Returns the ActionKind we will be performing with said action.

pub fn target(&self) -> TargetKind

Returns the OperationTarget we will be performing with said action.

pub fn into_action<C2: CrudOperation, O2: OperationTarget>(
self
) -> Action<S, C2, O2>

Changes both the CrudOperation and OperationTarget of this Action.

pub fn into_operation<O: CrudOperation>(self) -> Action<S, O, T>

Changes the CrudOperation of this Action.

pub fn into_target<T2: OperationTarget>(self) -> Action<S, C, T2>

Changes the OperationTarget of this Action.

pub fn into_create(self) -> Action<S, CreateOperation, T>

Sets the CrudOperation of this Action to CreateOperation.

pub fn into_read(self) -> Action<S, ReadOperation, T>

Sets the CrudOperation of this Action to ReadOperation.

pub fn into_update(self) -> Action<S, UpdateOperation, T>

Sets the CrudOperation of this Action to UpdateOperation.

pub fn into_delete(self) -> Action<S, DeleteOperation, T>

Sets the CrudOperation of this Action to DeleteOperation.

pub fn into_table(self) -> Action<S, C, TableTarget>

Sets the OperationTarget of this Action to TableTarget.

pub fn into_entry(self) -> Action<S, C, EntryTarget>

Sets the OperationTarget of this Action to EntryTarget.

pub fn with_entry<S2>(self) -> Action<S2, C, T>

Sets the target Entry of this Action.

pub fn set_table(&mut self, table_name: String) -> &mut Self

Sets the table for this action.

pub fn validate_table(&self) -> Result<(), ActionValidationError>

Validates that the table key is set.

Errors

Errors if Self::set_table has not yet been called.

pub fn validate_metadata(
&self,
key: Option<&str>
) -> Result<(), ActionValidationError>

This is supported on crate feature metadata only.

Validates that the key is not the private metadata key.

Errors

Errors if Self::set_key was passed the private metadata key.

impl<S: Entry, C: CrudOperation> Action<S, C, EntryTarget>

pub fn set_key<K: Key>(&mut self, key: &K) -> &mut Self

Sets the key for the action.

Users should prefer to call Self::set_entry over this, as setting the entry will automatically call this.

This is unused on TargetKind::Table actions.

pub fn set_data(&mut self, entity: &S) -> &mut Self

Sets the data for the action.

This is unused on TargetKind::Table actions.

pub fn validate_key(&self) -> Result<(), ActionValidationError>

Validate that the key has been set.

Errors

Errors if Self::set_key has not yet been called.

pub fn validate_data(&self) -> Result<(), ActionValidationError>

Validates that the data has been set.

Errors

Errors if Self::set_data has not yet been called.

pub fn validate_entry(&self) -> Result<(), ActionValidationError>

Validates that both the key and data have been set.

Errors

This errors if both the Self::set_key and Self::set_data (or Self::set_entry) has not been called.

impl<S: Entry> Action<S, CreateOperation, TableTarget>

pub fn create_table() -> Self

Creates a new CreateOperation based TableTarget operation.

impl<S: Entry> Action<S, ReadOperation, TableTarget>

pub fn read_table() -> Self

Creates a new ReadOperation based TableTarget operation.

impl<S: Entry> Action<S, DeleteOperation, TableTarget>

pub fn delete_table() -> Self

Creates a new DeleteOperation based TableTarget operation.

impl<S: Entry> Action<S, CreateOperation, EntryTarget>

pub fn create_entry() -> Self

Creates a new CreateOperation based EntryTarget operation.

impl<S: Entry> Action<S, ReadOperation, EntryTarget>

pub fn read_entry() -> Self

Creates a new ReadOperation based EntryTarget operation.

impl<S: Entry> Action<S, UpdateOperation, EntryTarget>

pub fn update_entry() -> Self

Creates a new UpdateOperation based EntryTarget operation.

impl<S: Entry> Action<S, DeleteOperation, EntryTarget>

pub fn delete_entry() -> Self

Creates a new DeleteOperation based EntryTarget operation.

impl<S: IndexEntry, C: CrudOperation> Action<S, C, EntryTarget>

pub fn set_entry(&mut self, entity: &S) -> &mut Self

Sets the Entry and Key that this Action will act over.

impl<S: Entry, C: CrudOperation, T: OperationTarget> Action<S, C, T>

pub async fn run<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<ActionResult<S>, ActionError<<B as Backend>::Error>>

Runs an Action to completion.

This method will dispatch to whatever method is needed for the action, and shouldn’t be used directly if it can be helped.

Errors

This will error if any of the other run prefixed methods cause an error.

Panics

This method will panic if CrudOperation is UpdateOperation and OperationTarget is TableTarget, as updating tables is not currently supported.

impl<S: Entry> Action<S, CreateOperation, EntryTarget>

pub async fn run_create_entry<B: Backend>(
self,
chart: &Starchart<B>
) -> Result<(), ActionError<<B as Backend>::Error>>

Validates and runs a CreateEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_entry fails, or if any of the Backend methods fail.

pub async unsafe fn run_create_entry_unchecked<B: Backend>(
self,
backend: &B
) -> Result<(), ActionRunError<<B as Backend>::Error>>

Runs a CreateEntryAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table, Self::set_data, and Self::set_key (or Self::set_entry for the last two) will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, ReadOperation, EntryTarget>

pub async fn run_read_entry<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<Option<S>, ActionError<<B as Backend>::Error>>

Validates and runs a ReadEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_key fails, or if any of the Backend methods fail.

pub async unsafe fn run_read_entry_unchecked<B: Backend>(
self,
backend: &B
) -> Result<Option<S>, ActionRunError<<B as Backend>::Error>>

Runs a ReadEntryAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table and Self::set_key will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, UpdateOperation, EntryTarget>

pub async fn run_update_entry<B: Backend>(
self,
chart: &Starchart<B>
) -> Result<(), ActionError<<B as Backend>::Error>>

Validates and runs a UpdateEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_entry fails, or if any of the Backend methods fail.

pub async unsafe fn run_update_entry_unchecked<B: Backend>(
self,
backend: &B
) -> Result<(), ActionRunError<<B as Backend>::Error>>

Runs a UpdateEntryAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table, Self::set_data, and Self::set_key (or Self::set_entry for the last two) will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, DeleteOperation, EntryTarget>

pub async fn run_delete_entry<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<bool, ActionError<<B as Backend>::Error>>

Validates and runs a DeleteEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_key fails, or if any of the Backend methods fail.

pub async unsafe fn run_delete_entry_unchecked<B: Backend>(
self,
backend: &B
) -> Result<bool, ActionRunError<<B as Backend>::Error>>

Runs a DeleteEntryAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table and Self::set_key will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, CreateOperation, TableTarget>

pub async fn run_create_table<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<(), ActionError<<B as Backend>::Error>>

Validates and runs a CreateTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

pub async unsafe fn run_create_table_unchecked<B: Backend>(
self,
backend: &B
) -> Result<(), ActionRunError<<B as Backend>::Error>>

Runs a CreateTableAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, ReadOperation, TableTarget>

pub async fn run_read_table<B: Backend, I>(
self,
gateway: &Starchart<B>
) -> Result<I, ActionError<<B as Backend>::Error>> where
I: FromIterator<S>,

Validates and runs a ReadTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

pub async unsafe fn run_read_table_unchecked<B: Backend, I>(
self,
backend: &B
) -> Result<I, ActionRunError<<B as Backend>::Error>> where
I: FromIterator<S>,

Runs a ReadTableAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

impl<S: Entry> Action<S, DeleteOperation, TableTarget>

pub async fn run_delete_table<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<bool, ActionError<<B as Backend>::Error>>

Validates and runs a DeleteTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

pub async unsafe fn run_delete_table_unchecked<B: Backend>(
self,
backend: &B
) -> Result<bool, ActionRunError<<B as Backend>::Error>>

Runs a DeleteTableAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.