Trait starchart::action::ActionRunner

pub trait ActionRunner<Success, Failure>: Sealed + Send {
    unsafe fn run<'a, B: Backend>(
        self, 
        gateway: &'a Gateway<B>
    ) -> Pin<Box<dyn Future<Output = Result<Success, Failure>> + Send + 'a>>
    where
        Failure: From<<B as Backend>::Error>;

    fn validate(&self) -> Result<(), ActionValidationError>;
}

The marker trait for all action runs, this trait should not be used and is only used to make the return type of Gateway::run easily known.

See Actions trait-implementations for more information.

This trait is sealed and cannot be implemented outside of this crate.

Required methods

unsafe fn run<'a, B: Backend>(
self,
gateway: &'a Gateway<B>
) -> Pin<Box<dyn Future<Output = Result<Success, Failure>> + Send + 'a>> where
Failure: From<<B as Backend>::Error>,

Runs the action through the Gateway.

Safety

This method may call a number of unsafe methods, such as Result::unwrap_unchecked and Option::unwrap_unchecked.

However, the Action is guaranteed to be safe to run if ActionRunner::validate is called beforehand, as any issues found will be reported before.

fn validate(&self) -> Result<(), ActionValidationError>

Validates that the Action has been created correctly.

Each individual implementation of this is specialized, for example, creating a table doesn’t have to check for a valid key to have been set.

If calling Self::run manually, this should be called first to avoid any unwanted behavior when performing database operations.

Errors

Any type of ActionValidationError that can arise.