Type Definition starchart::action::ReadEntryAction

pub type ReadEntryAction<S> = Action<S, ReadOperation, EntryTarget>;

A type alias for an Action with ReadOperation and EntryTarget as the parameters.

Implementations

impl<S: Entry> ReadEntryAction<S>

pub fn read_entry() -> Self

Creates a new ReadOperation based EntryTarget operation.

impl<S: Entry> ReadEntryAction<S>

pub async fn run_read_entry<B: Backend>(
self,
gateway: &Starchart<B>
) -> Result<Option<S>, ActionError>

Validates and runs a ReadEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_key fails, or if any of the Backend methods fail.

pub async unsafe fn run_read_entry_unchecked<B: Backend>(
self,
backend: &B
) -> Result<Option<S>, ActionRunError>

Runs a ReadEntryAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table and Self::set_key will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.