Type Definition starchart::action::ReadTableAction

pub type ReadTableAction<S> = Action<S, ReadOperation, TableTarget>;

A type alias for an Action with ReadOperation and TableTarget as the parameters.

Implementations

impl<S: Entry> ReadTableAction<S>

pub fn read_table() -> Self

Creates a new ReadOperation based TableTarget operation.

impl<S: Entry> ReadTableAction<S>

pub async fn run_read_table<B: Backend, I>(
self,
gateway: &Starchart<B>
) -> Result<I, ActionError> where
I: FromIterator<S>,

Validates and runs a ReadTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

Examples found in repository

src/action/mod.rs (line 505)

↕

464
465
466
467
468
469
470
471
472
473
474
475
476
477
478
479
480
481
482
483
484
485
486
487
488
489
490
491
492
493
494
495
496
497
498
499
500
501
502
503
504
505
506
507
508
509
510
511
512
513
514
515
516
517

pub async fn run<B: Backend>(
		self,
		gateway: &Starchart<B>,
	) -> Result<ActionResult<S>, ActionError> {
		match self.target() {
			TargetKind::Entry => match self.kind() {
				ActionKind::Create => {
					let create_entry: CreateEntryAction<S> = self.into_action();

					create_entry.run_create_entry(gateway).await?;
					Ok(ActionResult::Create)
				}
				ActionKind::Read => {
					let read_entry: ReadEntryAction<S> = self.into_action();

					let res = read_entry.run_read_entry(gateway).await?;
					Ok(ActionResult::SingleRead(res))
				}
				ActionKind::Update => {
					let update_entry: UpdateEntryAction<S> = self.into_action();

					update_entry.run_update_entry(gateway).await?;
					Ok(ActionResult::Update)
				}
				ActionKind::Delete => {
					let delete_entry: DeleteEntryAction<S> = self.into_action();

					let res = delete_entry.run_delete_entry(gateway).await?;
					Ok(ActionResult::Delete(res))
				}
			},
			TargetKind::Table => match self.kind() {
				ActionKind::Create => {
					let create_table: CreateTableAction<S> = self.into_action();

					create_table.run_create_table(gateway).await?;
					Ok(ActionResult::Create)
				}
				ActionKind::Read => {
					let read_table: ReadTableAction<S> = self.into_action();

					let res = read_table.run_read_table(gateway).await?;
					Ok(ActionResult::MultiRead(res))
				}
				ActionKind::Update => panic!("updating tables is unsupported"),
				ActionKind::Delete => {
					let delete_table: DeleteTableAction<S> = self.into_action();

					let res = delete_table.run_delete_table(gateway).await?;
					Ok(ActionResult::Delete(res))
				}
			},
		}
	}

pub async unsafe fn run_read_table_unchecked<B: Backend, I>(
self,
backend: &B
) -> Result<I, ActionRunError> where
I: FromIterator<S>,

Runs a ReadTableAction, not checking for any validation issues beforehand.

Errors

This will fail if any of the Backend methods fail.

Safety

This does not check any variants beforehand. Calling this without calling Self::set_table will cause immediate UB.

Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.

Examples found in repository

src/action/mod.rs (line 819)

↕

810
811
812
813
814
815
816
817
818
819
820
821
822

pub async fn run_read_table<B: Backend, I>(
		self,
		gateway: &Starchart<B>,
	) -> Result<I, ActionError>
	where
		I: FromIterator<S>,
	{
		self.validate_table()?;
		let lock = gateway.guard.shared();
		let res = unsafe { self.run_read_table_unchecked(gateway.backend()).await? };
		drop(lock);
		Ok(res)
	}