Module starchart::backend

The backend that fetches and provides data for the Starchart.

Modules

fsfs

Helpers for creating file-system based backends.

futures

Futures for Backend functions to return, for easier documentation.

Structs

BincodeBackendbincode

A Binary format based backend.

JsonBackendjson

A JSON based backend.

JsonPrettyBackendjson and pretty

A JSON based pretty printing backend.

MemoryBackendmemory

A memory-based backend, uses a DashMap of Values to represent data.

TomlBackendtoml

A TOML based backend.

TomlPrettyBackendtoml and pretty

A TOML based backend, that uses pretty printing.

YamlBackendyaml

A YAML based backend.

Traits

Backend

The backend to be used to manage data.