Trait starchart::backend::fs::FsBackend

pub trait FsBackend: Send + Sync {
    const EXTENSION: &'static str;

    fn from_reader<R, T>(&self, rdr: R) -> Result<T, FsError>
    where
        R: Read,
        T: Entry;

    fn to_bytes<T>(&self, value: &T) -> Result<Vec<u8>, FsError>
    where
        T: Entry;

    fn base_directory(&self) -> PathBuf;

    fn filename(id: &str) -> String { ... }

    fn resolve_path<P: AsRef<Path>>(&self, path: &[P]) -> PathBuf { ... }

    fn resolve_key<P: AsRef<Path>>(&self, p: P) -> Result<String, FsError> { ... }
}

This is supported on crate feature fs only.

The trait for all File System based backends to implement

This makes it easier to implement different file-system based databases.

Associated Constants

const EXTENSION: &'static str

The base extension of the files.

Required methods

fn from_reader<R, T>(&self, rdr: R) -> Result<T, FsError> where
R: Read,
T: Entry,

Turn a Reader into an Entry.

Errors

Implementors should return an error with FsError::serialization error upon failure.

fn to_bytes<T>(&self, value: &T) -> Result<Vec<u8>, FsError> where
T: Entry,

Turn a Entry into a writable Vec.

Errors

Implementors should return an error with FsError::deserialization error upon failure.

fn base_directory(&self) -> PathBuf

The base directory of the Fs database.

Provided methods

fn filename(id: &str) -> String

Turns a key into a valid filename, with the extension.

Examples found in repository

src/backend/fs.rs (line 328)

≺ ≻↕

323
324
325
326
327
328
329
330
331
332
333
334
335
336
337
338
339
340
341
342
343
344
345
346
347
348
349
350
351
352
353
354
355
356
357
358
359
360
361
362
363
364
365
366
367
368
369
370
371
372
373
374
375
376
377
378
379
380
381
382
383
384
385
386
387
388
389
390
391
392
393
394
395
396
397
398
399
400
401
402
403
404
405
406
407
408
409
410
411
412
413
414
415
416
417
418
419
420
421

fn get<'a, D>(&'a self, table: &'a str, id: &'a str) -> GetFuture<'a, D, Self::Error>
	where
		D: Entry,
	{
		Box::pin(async move {
			let filename = RW::filename(id);
			let path = self.resolve_path(&[table, filename.as_str()]);
			handle_io_result!(fs::File::open(&path).await, file, {
				let reader = io::BufReader::<StdFile>::new(file.into_std().await);
				Ok(Some(self.from_reader(reader)?))
			})
		})
	}

	fn has<'a>(&'a self, table: &'a str, id: &'a str) -> HasFuture<'a, Self::Error> {
		Box::pin(async move {
			let filename = RW::filename(id);
			let path = self.resolve_path(&[table, filename.as_str()]);
			handle_io_result!(fs::metadata(&path).await, _val, Ok(true), Ok(false))
		})
	}

	fn create<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> CreateFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			let filepath = RW::filename(id);

			let path = self.resolve_path(&[table, filepath.as_str()]);

			if self.has(table, id).await? {
				return Err(FsError {
					kind: FsErrorType::FileAlreadyExists { path },
					source: None,
				});
			}

			let serialized = self.to_bytes(value)?;

			fs::write(path, serialized).await.map_err(FsError::io)?;

			Ok(())
		})
	}

	fn update<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> UpdateFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			let serialized = self.to_bytes(value)?;
			let filepath = RW::filename(id);

			let path = self.resolve_path(&[table, filepath.as_str()]);

			fs::write(path, serialized).await.map_err(FsError::io)?;

			Ok(())
		})
	}

	fn replace<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> ReplaceFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			self.update(table, id, value).await?;

			Ok(())
		})
	}

	fn delete<'a>(&'a self, table: &'a str, id: &'a str) -> DeleteFuture<'a, Self::Error> {
		Box::pin(async move {
			let filename = RW::filename(id);

			fs::remove_file(self.resolve_path(&[table, filename.as_str()]))
				.await
				.map_err(FsError::io)?;

			Ok(())
		})
	}

fn resolve_path<P: AsRef<Path>>(&self, path: &[P]) -> PathBuf

Resolves the path of a file.

Examples found in repository

src/backend/fs.rs (line 263)

≺ ≻↕

261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326
327
328
329
330
331
332
333
334
335
336
337
338
339
340
341
342
343
344
345
346
347
348
349
350
351
352
353
354
355
356
357
358
359
360
361
362
363
364
365
366
367
368
369
370
371
372
373
374
375
376
377
378
379
380
381
382
383
384
385
386
387
388
389
390
391
392
393
394
395
396
397
398
399
400
401
402
403
404
405
406
407
408
409
410
411
412
413
414
415
416
417
418
419
420
421

fn has_table<'a>(&'a self, table: &'a str) -> HasTableFuture<'a, Self::Error> {
		Box::pin(async move {
			let result = fs::read_dir(self.resolve_path(&[table])).await;

			handle_io_result!(result, _val, Ok(true), Ok(false))
		})
	}

	fn create_table<'a>(&'a self, table: &'a str) -> CreateTableFuture<'a, Self::Error> {
		Box::pin(async move {
			fs::create_dir(self.resolve_path(&[table]))
				.await
				.map_err(FsError::io)?;

			Ok(())
		})
	}

	fn delete_table<'a>(&'a self, table: &'a str) -> DeleteTableFuture<'a, Self::Error> {
		Box::pin(async move {
			if self.has_table(table).await? {
				fs::remove_dir(self.resolve_path(&[table]))
					.await
					.map_err(FsError::io)?;
			}

			Ok(())
		})
	}

	fn get_keys<'a, I>(&'a self, table: &'a str) -> GetKeysFuture<'a, I, Self::Error>
	where
		I: FromIterator<String>,
	{
		Box::pin(async move {
			let mut stream = ReadDirStream::new(
				fs::read_dir(self.resolve_path(&[table]))
					.await
					.map_err(FsError::io)?,
			);
			let mut output = Vec::new();

			while let Some(raw) = stream.next().await {
				let entry = raw.map_err(FsError::io)?;

				if entry.file_type().await.map_err(FsError::io)?.is_dir() {
					continue; // coverage:ignore-line
				}

				let filename = self.resolve_key(entry.file_name()).ok();

				if filename.is_none() {
					continue; // coverage:ignore-line
				}

				output.push(unsafe { filename.inner_unwrap() });
			}

			Ok(output.into_iter().collect())
		})
	}

	fn get<'a, D>(&'a self, table: &'a str, id: &'a str) -> GetFuture<'a, D, Self::Error>
	where
		D: Entry,
	{
		Box::pin(async move {
			let filename = RW::filename(id);
			let path = self.resolve_path(&[table, filename.as_str()]);
			handle_io_result!(fs::File::open(&path).await, file, {
				let reader = io::BufReader::<StdFile>::new(file.into_std().await);
				Ok(Some(self.from_reader(reader)?))
			})
		})
	}

	fn has<'a>(&'a self, table: &'a str, id: &'a str) -> HasFuture<'a, Self::Error> {
		Box::pin(async move {
			let filename = RW::filename(id);
			let path = self.resolve_path(&[table, filename.as_str()]);
			handle_io_result!(fs::metadata(&path).await, _val, Ok(true), Ok(false))
		})
	}

	fn create<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> CreateFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			let filepath = RW::filename(id);

			let path = self.resolve_path(&[table, filepath.as_str()]);

			if self.has(table, id).await? {
				return Err(FsError {
					kind: FsErrorType::FileAlreadyExists { path },
					source: None,
				});
			}

			let serialized = self.to_bytes(value)?;

			fs::write(path, serialized).await.map_err(FsError::io)?;

			Ok(())
		})
	}

	fn update<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> UpdateFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			let serialized = self.to_bytes(value)?;
			let filepath = RW::filename(id);

			let path = self.resolve_path(&[table, filepath.as_str()]);

			fs::write(path, serialized).await.map_err(FsError::io)?;

			Ok(())
		})
	}

	fn replace<'a, S>(
		&'a self,
		table: &'a str,
		id: &'a str,
		value: &'a S,
	) -> ReplaceFuture<'a, Self::Error>
	where
		S: Entry,
	{
		Box::pin(async move {
			self.update(table, id, value).await?;

			Ok(())
		})
	}

	fn delete<'a>(&'a self, table: &'a str, id: &'a str) -> DeleteFuture<'a, Self::Error> {
		Box::pin(async move {
			let filename = RW::filename(id);

			fs::remove_file(self.resolve_path(&[table, filename.as_str()]))
				.await
				.map_err(FsError::io)?;

			Ok(())
		})
	}

fn resolve_key<P: AsRef<Path>>(&self, p: P) -> Result<String, FsError>

Turns a filename into a valid key.

Errors

Returns an FsError when the file path does not have the correct extension.

Examples found in repository

src/backend/fs.rs (line 310)

↕

291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321

fn get_keys<'a, I>(&'a self, table: &'a str) -> GetKeysFuture<'a, I, Self::Error>
	where
		I: FromIterator<String>,
	{
		Box::pin(async move {
			let mut stream = ReadDirStream::new(
				fs::read_dir(self.resolve_path(&[table]))
					.await
					.map_err(FsError::io)?,
			);
			let mut output = Vec::new();

			while let Some(raw) = stream.next().await {
				let entry = raw.map_err(FsError::io)?;

				if entry.file_type().await.map_err(FsError::io)?.is_dir() {
					continue; // coverage:ignore-line
				}

				let filename = self.resolve_key(entry.file_name()).ok();

				if filename.is_none() {
					continue; // coverage:ignore-line
				}

				output.push(unsafe { filename.inner_unwrap() });
			}

			Ok(output.into_iter().collect())
		})
	}