Trait starchart::backend::Backend

pub trait Backend: Send + Sync {
    type Error: Send + Sync + StdError + 'static;

    fn has_table<'a>(
        &'a self, 
        table: &'a str
    ) -> HasTableFuture<'a, Self::Error>;

    fn create_table<'a>(
        &'a self, 
        table: &'a str
    ) -> CreateTableFuture<'a, Self::Error>;

    fn delete_table<'a>(
        &'a self, 
        table: &'a str
    ) -> DeleteTableFuture<'a, Self::Error>;

    fn get_keys<'a, I>(
        &'a self, 
        table: &'a str
    ) -> GetKeysFuture<'a, I, Self::Error>
    where
        I: FromIterator<String>;

    fn get<'a, D>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> GetFuture<'a, D, Self::Error>
    where
        D: Entry;

    fn has<'a>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> HasFuture<'a, Self::Error>;

    fn create<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> CreateFuture<'a, Self::Error>
    where
        S: Entry;

    fn update<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> UpdateFuture<'a, Self::Error>
    where
        S: Entry;

    fn replace<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> ReplaceFuture<'a, Self::Error>
    where
        S: Entry;

    fn delete<'a>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> DeleteFuture<'a, Self::Error>;

    fn init(&self) -> InitFuture<'_, Self::Error> { ... }

    unsafe fn shutdown(&self) -> ShutdownFuture<'_> { ... }

    fn ensure_table<'a>(
        &'a self, 
        table: &'a str
    ) -> EnsureTableFuture<'a, Self::Error> { ... }

    fn get_all<'a, D, I>(
        &'a self, 
        table: &'a str, 
        entries: &'a [&str]
    ) -> GetAllFuture<'a, I, Self::Error>
    where
        D: Entry,
        I: FromIterator<D>,
    { ... }

    fn ensure<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> EnsureFuture<'a, Self::Error>
    where
        S: Entry,
    { ... }

}

The backend to be used to manage data.

Associated Types

type Error: Send + Sync + StdError + 'static

The Error type that the backend will report up.

Required methods

fn has_table<'a>(&'a self, table: &'a str) -> HasTableFuture<'a, Self::Error>

Check if a table exists.

fn create_table<'a>(
&'a self,
table: &'a str
) -> CreateTableFuture<'a, Self::Error>

Inserts or creates a table.

fn delete_table<'a>(
&'a self,
table: &'a str
) -> DeleteTableFuture<'a, Self::Error>

Deletes or drops a table.

fn get_keys<'a, I>(
&'a self,
table: &'a str
) -> GetKeysFuture<'a, I, Self::Error> where
I: FromIterator<String>,

Gets all the keys in the table.

fn get<'a, D>(
&'a self,
table: &'a str,
id: &'a str
) -> GetFuture<'a, D, Self::Error> where
D: Entry,

Gets a certain entry from a table.

fn has<'a>(&'a self, table: &'a str, id: &'a str) -> HasFuture<'a, Self::Error>

Checks if an entry exists in a table.

fn create<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> CreateFuture<'a, Self::Error> where
S: Entry,

Inserts a new entry into a table.

fn update<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> UpdateFuture<'a, Self::Error> where
S: Entry,

Updates an existing entry in a table.

fn replace<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> ReplaceFuture<'a, Self::Error> where
S: Entry,

Replaces an existing entry in a table.

fn delete<'a>(
&'a self,
table: &'a str,
id: &'a str
) -> DeleteFuture<'a, Self::Error>

Deletes an entry from a table.

Provided methods

fn init(&self) -> InitFuture<'_, Self::Error>

An optional initialization function, useful for making connections to the database.

The default impl does nothing

Examples found in repository

src/starchart.rs (line 38)

37
38
39
40
41
42
43

pub async fn new(backend: B) -> Result<Self, B::Error> {
		backend.init().await?;
		Ok(Self {
			backend: Arc::new(backend),
			guard: Arc::default(),
		})
	}

unsafe fn shutdown(&self) -> ShutdownFuture<'_>

An optional shutdown function, useful for disconnecting from databases gracefully.

The default impl does nothing

Safety

This should not fail, as it’s ran upon dropping the Starchart, and panicking during a drop means resources haven’t adequately been cleaned up, which isn’t inherintly UB however it should still be documented.

Examples found in repository

src/starchart.rs (line 57)

56
57
58

fn drop(&mut self) {
		block_on(unsafe { self.backend.shutdown() });
	}

fn ensure_table<'a>(
&'a self,
table: &'a str
) -> EnsureTableFuture<'a, Self::Error>

Ensures a table exists. Uses Self::has_table first, then Self::create_table if it returns false.

Examples found in repository

src/action/mod.rs (line 795)

↕

788
789
790
791
792
793
794
795
796
797
798
799
800
801
802
803
804
805
806
807
808
809
810
811
812
813
814
815

pub async unsafe fn run_create_table_unchecked<B: Backend>(
		self,
		backend: &B,
	) -> Result<(), ActionRunError> {
		let table = self.inner.table.inner_unwrap();

		backend
			.ensure_table(&table)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		#[cfg(feature = "metadata")]
		{
			let metadata = S::default();
			backend
				.ensure(&table, METADATA_KEY, &metadata)
				.await
				.map_err(|e| ActionRunError {
					source: Some(Box::new(e)),
					kind: ActionRunErrorType::Backend,
				})?;
		}

		Ok(())
	}

fn get_all<'a, D, I>(
&'a self,
table: &'a str,
entries: &'a [&str]
) -> GetAllFuture<'a, I, Self::Error> where
D: Entry,
I: FromIterator<D>,

Gets all entries that match a predicate, to get all entries, use get_keys first.

Examples found in repository

src/action/mod.rs (line 881)

↕

852
853
854
855
856
857
858
859
860
861
862
863
864
865
866
867
868
869
870
871
872
873
874
875
876
877
878
879
880
881
882
883
884
885
886
887
888
889

pub async unsafe fn run_read_table_unchecked<B: Backend, I>(
		mut self,
		backend: &B,
	) -> Result<I, ActionRunError>
	where
		I: FromIterator<S>,
	{
		let table = self.inner.table.take().inner_unwrap();

		#[cfg(feature = "metadata")]
		self.check_metadata(backend, &table).await?;

		let keys = backend
			.get_keys::<Vec<_>>(&table)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		#[cfg(feature = "metadata")]
		let keys = keys
			.into_iter()
			.filter(|value| value != METADATA_KEY)
			.collect::<Vec<_>>();

		let keys_borrowed = keys.iter().map(String::as_str).collect::<Vec<_>>();

		let data = backend
			.get_all::<S, I>(&table, &keys_borrowed)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		Ok(data)
	}

fn ensure<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> EnsureFuture<'a, Self::Error> where
S: Entry,

Ensures a value exists in the table.

Examples found in repository

src/action/mod.rs (line 806)

↕

788
789
790
791
792
793
794
795
796
797
798
799
800
801
802
803
804
805
806
807
808
809
810
811
812
813
814
815

pub async unsafe fn run_create_table_unchecked<B: Backend>(
		self,
		backend: &B,
	) -> Result<(), ActionRunError> {
		let table = self.inner.table.inner_unwrap();

		backend
			.ensure_table(&table)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		#[cfg(feature = "metadata")]
		{
			let metadata = S::default();
			backend
				.ensure(&table, METADATA_KEY, &metadata)
				.await
				.map_err(|e| ActionRunError {
					source: Some(Box::new(e)),
					kind: ActionRunErrorType::Backend,
				})?;
		}

		Ok(())
	}