Struct starchart::Starchart

pub struct Starchart<B: Backend> { /* private fields */ }

The base structure for managing data.

The inner data is wrapped in an Arc, so cloning is cheap and will allow multiple accesses to the data.

Implementations

impl<B: Backend> Starchart<B>

pub unsafe fn backend(&self) -> &B

Gives access to the raw Backend instance.

Safety

Accessing the backend functions directly isn’t inheritly unsafe, however care must be taken to ensure the data isn’t modified directly, and that Backend::shutdown isn’t directly called.

Examples found in repository

src/action/mod.rs (line 534)

≺ ≻↕

526
527
528
529
530
531
532
533
534
535
536
537
538
539
540
541
542
543
544
545
546
547
548
549
550
551
552
553
554
555
556
557
558
559
560
561
562
563
564
565
566
567
568
569
570
571
572
573
574
575
576
577
578
579
580
581
582
583
584
585
586
587
588
589
590
591
592
593
594
595
596
597
598
599
600
601
602
603
604
605
606
607
608
609
610
611
612
613
614
615
616
617
618
619
620
621
622
623
624
625
626
627
628
629
630
631
632
633
634
635
636
637
638
639
640
641
642
643
644
645
646
647
648
649
650
651
652
653
654
655
656
657
658
659
660
661
662
663
664
665
666
667
668
669
670
671
672
673
674
675
676
677
678
679
680
681
682
683
684
685
686
687
688
689
690
691
692
693
694
695
696
697
698
699
700
701
702
703
704
705
706
707
708
709
710
711
712
713
714
715
716
717
718
719
720
721
722
723
724
725
726
727
728
729
730
731
732
733
734
735
736
737
738
739
740
741
742
743
744
745
746
747
748
749
750
751
752
753
754
755
756
757
758
759
760
761
762
763
764
765
766
767
768
769
770
771
772
773
774
775
776
777
778
779
780
781
782
783
784
785
786
787
788
789
790
791
792
793
794
795
796
797
798
799
800
801
802
803
804
805
806
807
808
809
810
811
812
813
814
815
816
817
818
819
820
821
822
823
824
825
826
827
828
829
830
831
832
833
834
835
836
837
838
839
840
841
842
843
844
845
846
847
848
849
850
851
852
853
854
855
856
857
858
859
860
861
862
863
864
865
866
867
868
869
870
871
872
873
874
875
876
877
878
879
880
881
882
883
884
885
886
887
888
889
890
891
892
893
894
895
896
897
898
899
900
901
902
903
904
905
906
907
908
909
910
911

pub async fn run_create_entry<B: Backend>(
		self,
		chart: &Starchart<B>,
	) -> Result<(), ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		self.validate_entry().map_err(ActionError::validation)?;
		let lock = chart.guard.exclusive();
		unsafe {
			self.run_create_entry_unchecked(chart.backend())
				.await
				.map_err(ActionError::run)?;
		};
		drop(lock);
		Ok(())
	}

	/// Runs a [`CreateEntryAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`], [`Self::set_data`],
	/// and [`Self::set_key`] (or [`Self::set_entry`] for the last two) will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_create_entry_unchecked<B: Backend>(
		mut self,
		backend: &B,
	) -> Result<(), ActionRunError> {
		let table_name = self.inner.table.take().inner_unwrap();
		let key = self.inner.key.take().inner_unwrap();
		let entry = self.inner.data.take().inner_unwrap();
		if backend
			.has(&table_name, &key)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})? {
			return Ok(());
		}

		#[cfg(feature = "metadata")]
		self.check_metadata(backend, &table_name).await?;

		backend
			.create(&table_name, &key, &*entry)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		Ok(())
	}
}

impl<S: Entry> ReadEntryAction<S> {
	/// Validates and runs a [`ReadEntryAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] or [`Self::validate_key`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_read_entry<B: Backend>(
		self,
		gateway: &Starchart<B>,
	) -> Result<Option<S>, ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		self.validate_key().map_err(ActionError::validation)?;
		let lock = gateway.guard.exclusive();
		let res = unsafe {
			self.run_read_entry_unchecked(gateway.backend())
				.await
				.map_err(ActionError::run)?
		};
		drop(lock);
		Ok(res)
	}

	/// Runs a [`ReadEntryAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`] and [`Self::set_key`] will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_read_entry_unchecked<B: Backend>(
		mut self,
		backend: &B,
	) -> Result<Option<S>, ActionRunError> {
		let table_name = self.inner.table.take().inner_unwrap();
		let key = self.inner.key.take().inner_unwrap();

		#[cfg(feature = "metadata")]
		self.check_metadata(backend, &table_name).await?;

		backend
			.get(&table_name, &key)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})
	}
}

impl<S: Entry> UpdateEntryAction<S> {
	/// Validates and runs a [`UpdateEntryAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] or [`Self::validate_entry`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_update_entry<B: Backend>(
		self,
		chart: &Starchart<B>,
	) -> Result<(), ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		self.validate_entry().map_err(ActionError::validation)?;
		let lock = chart.guard.exclusive();
		unsafe {
			self.run_update_entry_unchecked(chart.backend())
				.await
				.map_err(ActionError::run)?;
		};
		drop(lock);
		Ok(())
	}

	/// Runs a [`UpdateEntryAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`], [`Self::set_data`],
	/// and [`Self::set_key`] (or [`Self::set_entry`] for the last two) will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_update_entry_unchecked<B: Backend>(
		mut self,
		backend: &B,
	) -> Result<(), ActionRunError> {
		let table = self.inner.table.take().inner_unwrap();
		let key = self.inner.key.take().inner_unwrap();
		let new_data = self.inner.data.take().inner_unwrap();

		#[cfg(feature = "metadata")]
		self.check_metadata(backend, &table).await?;

		backend
			.update(&table, &key, &new_data)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})
	}
}

impl<S: Entry> DeleteEntryAction<S> {
	/// Validates and runs a [`DeleteEntryAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] or [`Self::validate_key`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_delete_entry<B: Backend>(
		self,
		gateway: &Starchart<B>,
	) -> Result<bool, ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		self.validate_key().map_err(ActionError::validation)?;
		let lock = gateway.guard.exclusive();
		let res = unsafe {
			self.run_delete_entry_unchecked(gateway.backend())
				.await
				.map_err(ActionError::run)?
		};
		drop(lock);
		Ok(res)
	}

	/// Runs a [`DeleteEntryAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`] and [`Self::set_key`] will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_delete_entry_unchecked<B: Backend>(
		mut self,
		backend: &B,
	) -> Result<bool, ActionRunError> {
		let table = self.inner.table.take().inner_unwrap();
		let key = self.inner.key.take().inner_unwrap();
		let exists = backend
			.has(&table, &key)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		backend
			.delete(&table, &key)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		let after_exists = backend
			.has(&table, &key)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		Ok(exists != after_exists)
	}
}

impl<S: Entry> CreateTableAction<S> {
	/// Validates and runs a [`CreateTableAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_create_table<B: Backend>(
		self,
		gateway: &Starchart<B>,
	) -> Result<(), ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		let lock = gateway.guard.exclusive();
		unsafe {
			self.run_create_table_unchecked(gateway.backend())
				.await
				.map_err(ActionError::run)?;
		};
		drop(lock);
		Ok(())
	}

	/// Runs a [`CreateTableAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`] will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_create_table_unchecked<B: Backend>(
		self,
		backend: &B,
	) -> Result<(), ActionRunError> {
		let table = self.inner.table.inner_unwrap();

		backend
			.ensure_table(&table)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		#[cfg(feature = "metadata")]
		{
			let metadata = S::default();
			backend
				.ensure(&table, METADATA_KEY, &metadata)
				.await
				.map_err(|e| ActionRunError {
					source: Some(Box::new(e)),
					kind: ActionRunErrorType::Backend,
				})?;
		}

		Ok(())
	}
}

impl<S: Entry> ReadTableAction<S> {
	/// Validates and runs a [`ReadTableAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_read_table<B: Backend, I>(
		self,
		gateway: &Starchart<B>,
	) -> Result<I, ActionError>
	where
		I: FromIterator<S>,
	{
		self.validate_table().map_err(ActionError::validation)?;
		let lock = gateway.guard.shared();
		let res = unsafe {
			self.run_read_table_unchecked(gateway.backend())
				.await
				.map_err(ActionError::run)?
		};
		drop(lock);
		Ok(res)
	}

	/// Runs a [`ReadTableAction`], not checking for any validation issues beforehand.
	///
	/// # Errors
	///
	/// This will fail if any of the [`Backend`] methods fail.
	///
	/// # Safety
	/// This does not check any variants beforehand. Calling this without calling [`Self::set_table`] will cause immediate `UB`.
	///
	/// Additionally, this does no atomic locking, so running this at the same time as a read or write will cause data races.
	pub async unsafe fn run_read_table_unchecked<B: Backend, I>(
		mut self,
		backend: &B,
	) -> Result<I, ActionRunError>
	where
		I: FromIterator<S>,
	{
		let table = self.inner.table.take().inner_unwrap();

		#[cfg(feature = "metadata")]
		self.check_metadata(backend, &table).await?;

		let keys = backend
			.get_keys::<Vec<_>>(&table)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		#[cfg(feature = "metadata")]
		let keys = keys
			.into_iter()
			.filter(|value| value != METADATA_KEY)
			.collect::<Vec<_>>();

		let keys_borrowed = keys.iter().map(String::as_str).collect::<Vec<_>>();

		let data = backend
			.get_all::<S, I>(&table, &keys_borrowed)
			.await
			.map_err(|e| ActionRunError {
				source: Some(Box::new(e)),
				kind: ActionRunErrorType::Backend,
			})?;

		Ok(data)
	}
}

impl<S: Entry> DeleteTableAction<S> {
	/// Validates and runs a [`DeleteTableAction`].
	///
	/// # Errors
	///
	/// This returns an error if [`Self::validate_table`] fails, or if any of the [`Backend`] methods fail.
	pub async fn run_delete_table<B: Backend>(
		self,
		gateway: &Starchart<B>,
	) -> Result<bool, ActionError> {
		self.validate_table().map_err(ActionError::validation)?;
		let lock = gateway.guard.exclusive();
		let res = unsafe {
			self.run_delete_table_unchecked(gateway.backend())
				.await
				.map_err(ActionError::run)?
		};
		drop(lock);
		Ok(res)
	}

pub async fn new(backend: B) -> Result<Self, B::Error>

Creates a new Starchart, and initializes the Backend.

Errors

Any errors that Backend::init can raise.

Examples found in repository

examples/get_all.rs (line 44)

↕

43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101

fn main() -> Result<()> {
	let chart = Starchart::new(JsonBackend::new("./examples/get_all_db")?).await?;

	{
		let mut action: CreateTableAction<Settings> = Action::new();
		action.set_table("foo".to_owned());

		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	let mut read_all_action: ReadTableAction<Settings> = Action::new();
	read_all_action.set_table("foo".to_owned());

	// let mut values: Vec<Settings> = chart.run(read_all_action).await??;
	let mut values = read_all_action
		.run(&chart)
		.await?
		.unwrap_multi_read::<Vec<_>>();

	values.sort();

	assert_eq!(
		values,
		vec![
			Settings {
				id: 1,
				name: "John Doe".to_owned(),
				age: 21
			},
			Settings {
				id: 2,
				name: "Ferris".to_owned(),
				age: 42
			},
			Settings {
				id: 3,
				name: "The Queen".to_owned(),
				age: 73
			}
		]
	);

	Ok(())
}

examples/basic.rs (line 32)

↕

30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80

fn main() -> Result<()> {
	// Create an initialize the database.
	let chart = Starchart::new(MemoryBackend::default()).await?;

	// Create and run an action to create the "foo" table with the Settings struct.
	let mut create_table_action: CreateTableAction<Settings> = Action::new();

	create_table_action.set_table("foo".to_owned());

	// chart.run(create_table_action).await??;
	create_table_action.run(&chart).await?.unwrap_create();

	// Insert some entries into the table.
	for (age, name) in vec![
		(21, "John Doe".to_owned()),
		(42, "Ferris".to_owned()),
		(73, "The Queen".to_owned()),
	] {
		let mut action: CreateEntryAction<Settings> = Action::new();
		action
			.set_table("foo".to_owned())
			.set_entry(&Settings::new(name, age));
		// chart.run(action).await??;
		action.run(&chart).await?.unwrap_create();
	}

	// Get a single entry.

	let the_queen = {
		// Action type helpers are named after their CRUD counterparts; Create, Read, Update, and Delete.
		let mut action: ReadEntryAction<Settings> = Action::new();
		action.set_key(&3_u64).set_table("foo".to_owned());

		action
			.run(&chart)
			.await?
			.unwrap_single_read()
			.expect("the queen has fallen!")
	};

	assert_eq!(
		the_queen,
		Settings {
			id: 3,
			name: "The Queen".to_owned(),
			age: 73
		}
	);

	Ok(())
}