Enum starchart::action::ActionResult

pub enum ActionResult<R> {
    Create,
    SingleRead(Option<R>),
    MultiRead(Vec<R>),
    Update,
    Delete(bool),
}

A custom Result type that allows the Action::run method to return multiple different types.

Variants

Create

Indicates a CreateOperation was performed.

SingleRead(Option<R>)

Tuple Fields

0: Option<R>

Indicates a ReadOperation was performed on an EntryTarget.

This value can be retrieved with Self::unwrap_single_read.

MultiRead(Vec<R>)

Tuple Fields

0: Vec<R>

Indicates a ReadOperation was performed on a TableTarget.

This value can be retrieved with Self::unwrap_multi_read.

Update

Indicates an UpdateOperation was performed.

Delete(bool)

Tuple Fields

0: bool

Indicates a DeleteOperation was performed.

The inner value can be retrieved with Self::unwrap_delete to determine whether the delete was successful.

Implementations

impl<R> ActionResult<R>

pub const fn is_create(&self) -> bool

Returns true if the result is Create.

pub const fn is_single_read(&self) -> bool

Returns true if the result is SingleRead.

pub const fn is_multi_read(&self) -> bool

Returns true if the result is MultiRead.

pub const fn is_read(&self) -> bool

Returns true if the result is either a SingleRead or a MultiRead.

pub const fn is_update(&self) -> bool

Returns true if the result is an Update.

pub const fn is_delete(&self) -> bool

Returns true if the result is a Delete.

impl<R: Entry> ActionResult<R>

pub fn unwrap_create(self)

Unwraps the ActionResult and asserts it’s a create

Panics

This method panics if Self is anything other than Create.

pub unsafe fn unwrap_create_unchecked(self)

Unwraps the ActionResult::Create value, without checking if the value is any other type.

Safety

Calling this method on anything other than a Create will cause UB.

pub fn unwrap_single_read(self) -> Option<R>

Unwraps the ActionResult and asserts it’s a single read.

Panics

This method panics if Self is anything other than SingleRead.

pub unsafe fn unwrap_single_read_unchecked(self) -> Option<R>

Unwraps the ActionResult::SingleRead value, without checking if the value is any other type.

Safety

Calling this method on anything other than a SingleRead will cause UB.

pub fn unwrap_multi_read<I: FromIterator<R>>(self) -> I

Unwraps the ActionResult and asserts it’s a multi-read.

Panics

This method panics if Self is anything other than MultiRead.

pub unsafe fn unwrap_multi_read_unchecked<I: FromIterator<R>>(self) -> I

Unwraps the ActionResult::MultiRead value, without checking if the value is any other type.

Safety

Calling this method on anything other than a MultiRead will cause UB.

pub fn unwrap_update(self)

Unwraps the ActionResult and asserts it’s an update.

Panics

This method panics if Self is anything other than Update.

pub unsafe fn unwrap_update_unchecked(self)

Unwraps the ActionResult::Update value, without checking if the value is any other type.

Safety

Calling this method on anything other than an Update will cause UB.

pub fn unwrap_delete(self) -> bool

Unwraps the ActionResult and asserts it’s a delete.

Panics

This method panics if Self is anything other than Delete.

pub unsafe fn unwrap_delete_unchecked(self) -> bool

Unwraps the ActionResult::Delete value, without checking if the value is any other type.

Safety

Calling this method on anything other than a Delete will cause UB.