Struct starchart::action::Action

pub struct Action<'a, S, C, T> { /* private fields */ }

An Action for easy CRUD operations within a Starchart.

Implementations

impl<'a, S, C, T> Action<'a, S, C, T>

pub const fn new() -> Self

Creates a new Action with the specified operation.

pub const fn table(&self) -> Option<&str>

Get a reference to the currently set table.

pub fn key(&self) -> Option<&str>

Get a reference to the currently set key.

impl<'a, S: Entry, C: CrudOperation, T: OperationTarget> Action<'a, S, C, T>

pub fn data(&self) -> Option<&S>

Get a reference to the currently set data.

pub fn kind(&self) -> ActionKind

Returns the ActionKind we will be performing with said action.

pub fn target(&self) -> TargetKind

Returns the OperationTarget we will be performing with said action.

pub fn to_dynamic(&self) -> DynamicAction<S>

Converts the action to a dynamic one, usually for serialization of some sort.

pub fn set_table(&mut self, table_name: &'a str) -> &mut Self

Sets the table for this action.

pub fn validate_table(&self) -> Result<(), ActionValidationError>

Validates that the table key is set.

Errors

Errors if Self::set_table has not yet been called.

pub fn validate_metadata(
&self,
key: Option<&str>
) -> Result<(), ActionValidationError>

This is supported on crate feature metadata only.

Validates that the key is not the private metadata key.

Errors

Errors if Self::set_key was passed the private metadata key.

impl<'a, S: Entry, C: CrudOperation> Action<'a, S, C, EntryTarget>

pub fn set_key<K: Key>(&mut self, key: &K) -> &mut Self

Sets the key for the action.

Users should prefer to call Self::set_entry over this, as setting the entry will automatically call this.

This is unused on TargetKind::Table actions.

pub fn set_data(&mut self, entity: &'a S) -> &mut Self

Sets the data for the action.

This is unused on TargetKind::Table actions.

pub fn validate_key(&self) -> Result<(), ActionValidationError>

Validate that the key has been set.

Errors

Errors if Self::set_key has not yet been called.

pub fn validate_data(&self) -> Result<(), ActionValidationError>

Validates that the data has been set.

Errors

Errors if Self::set_data has not yet been called.

pub fn validate_entry(&self) -> Result<(), ActionValidationError>

Validates that both the key and data have been set.

Errors

This errors if both the Self::set_key and Self::set_data (or Self::set_entry) has not been called.

impl<'a, S: IndexEntry, C: CrudOperation> Action<'a, S, C, EntryTarget>

pub fn set_entry(&mut self, entity: &'a S) -> &mut Self

Sets the Entry and Key that this Action will act over.

impl<'a, S: Entry> Action<'a, S, CreateOperation, EntryTarget>

pub fn run_create_entry<B: Backend>(
self,
chart: &'a Starchart<B>
) -> impl Future<Output = Result<(), ActionError>> + 'a

Validates and runs a CreateEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_entry fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, ReadOperation, EntryTarget>

pub fn run_read_entry<B: Backend>(
self,
gateway: &'a Starchart<B>
) -> impl Future<Output = Result<Option<S>, ActionError>> + 'a

Validates and runs a ReadEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_key fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, UpdateOperation, EntryTarget>

pub fn run_update_entry<B: Backend>(
self,
chart: &'a Starchart<B>
) -> impl Future<Output = Result<(), ActionError>> + 'a

Validates and runs a UpdateEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_entry fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, DeleteOperation, EntryTarget>

pub fn run_delete_entry<B: Backend>(
self,
gateway: &'a Starchart<B>
) -> impl Future<Output = Result<bool, ActionError>> + 'a

Validates and runs a DeleteEntryAction.

Errors

This returns an error if Self::validate_table or Self::validate_key fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, CreateOperation, TableTarget>

pub fn run_create_table<B: Backend>(
self,
gateway: &'a Starchart<B>
) -> impl Future<Output = Result<(), ActionError>> + 'a

Validates and runs a CreateTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, ReadOperation, TableTarget>

pub fn run_read_table<B: Backend, I>(
self,
gateway: &'a Starchart<B>
) -> impl Future<Output = Result<I, ActionError>> + 'a where
I: FromIterator<S> + 'a,

Validates and runs a ReadTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.

impl<'a, S: Entry> Action<'a, S, DeleteOperation, TableTarget>

pub fn run_delete_table<B: Backend>(
self,
gateway: &'a Starchart<B>
) -> impl Future<Output = Result<bool, ActionError>> + 'a

Validates and runs a DeleteTableAction.

Errors

This returns an error if Self::validate_table fails, or if any of the Backend methods fail.