Module starchart::backend

The backend that fetches and provides data for the Starchart.

Modules

fsfs

Helpers for creating file-system based backends.

future

Futures for Backend functions to return, for easier documentation.

Structs

BincodeBackendbincode

A Binary format based backend.

JsonBackendjson

A JSON based backend.

MemoryBackendmemory

A memory-based backend, uses a [DashMap] of Values to represent data.

RonBackendron

A RON based backend.

TomlBackendtoml

A TOML based backend.

YamlBackendyaml

A YAML based backend.

Traits

Backend

The backend to be used to manage data.