Trait starchart::backend::fs::FsBackend

pub trait FsBackend: Send + Sync {
    const EXTENSION: &'static str;

    fn from_reader<R, T>(rdr: R) -> Result<T, FsError>
    where
        R: Read,
        T: Entry;

    fn to_bytes<T>(value: &T) -> Result<Vec<u8>, FsError>
    where
        T: Entry;

    fn base_directory(&self) -> PathBuf;

    fn filename(id: &str) -> String { ... }

    fn resolve_path<P: AsRef<Path>>(&self, path: &[P]) -> PathBuf { ... }

    fn resolve_key<P: AsRef<Path>>(&self, p: P) -> Result<String, FsError> { ... }
}

This is supported on crate feature fs only.

The trait for all File System based backends to implement

This makes it easier to implement different file-system based databases.

Associated Constants

const EXTENSION: &'static str

The base extension of the files.

Required methods

fn from_reader<R, T>(rdr: R) -> Result<T, FsError> where
R: Read,
T: Entry,

Turn a Reader into an Entry.

Errors

Implementors should return a FsError::Serde error upon failure.

fn to_bytes<T>(value: &T) -> Result<Vec<u8>, FsError> where
T: Entry,

Turn a Entry into a writable Vec.

Errors

Implementors should return a FsError::Serde error upon failure.

fn base_directory(&self) -> PathBuf

The base directory of the Fs database.

Provided methods

fn filename(id: &str) -> String

Turns a key into a valid filename, with the extension.

fn resolve_path<P: AsRef<Path>>(&self, path: &[P]) -> PathBuf

Resolves the path of a file.

fn resolve_key<P: AsRef<Path>>(&self, p: P) -> Result<String, FsError>

Turns a filename into a valid key.

Errors

Returns a FsError::InvalidFile when the file path does not have the correct extension.