Trait starchart::backend::Backend

pub trait Backend: Send + Sync {
    type Error: Send + Sync + StdError;

    fn has_table<'a>(
        &'a self, 
        table: &'a str
    ) -> HasTableFuture<'a, Self::Error>;

    fn create_table<'a>(
        &'a self, 
        table: &'a str
    ) -> CreateTableFuture<'a, Self::Error>;

    fn delete_table<'a>(
        &'a self, 
        table: &'a str
    ) -> DeleteTableFuture<'a, Self::Error>;

    fn get_keys<'a, I>(
        &'a self, 
        table: &'a str
    ) -> GetKeysFuture<'a, I, Self::Error>
    where
        I: FromIterator<String>;

    fn get<'a, D>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> GetFuture<'a, D, Self::Error>
    where
        D: for<'de> Deserialize<'de> + Send + Sync;

    fn has<'a>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> HasFuture<'a, Self::Error>;

    fn create<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> CreateFuture<'a, Self::Error>
    where
        S: Serialize + Send + Sync;

    fn update<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> UpdateFuture<'a, Self::Error>
    where
        S: Serialize + Send + Sync;

    fn replace<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> ReplaceFuture<'a, Self::Error>
    where
        S: Serialize + Send + Sync;

    fn delete<'a>(
        &'a self, 
        table: &'a str, 
        id: &'a str
    ) -> DeleteFuture<'a, Self::Error>;

    fn init(&self) -> InitFuture<'_, Self::Error> { ... }

    unsafe fn shutdown(&self) -> ShutdownFuture<'_> { ... }

    fn ensure_table<'a>(
        &'a self, 
        table: &'a str
    ) -> EnsureTableFuture<'a, Self::Error> { ... }

    fn get_all<'a, D, I>(
        &'a self, 
        table: &'a str, 
        entries: &'a [&str]
    ) -> GetAllFuture<'a, I, Self::Error>
    where
        D: for<'de> Deserialize<'de> + Send + Sync,
        I: FromIterator<D>,
    { ... }

    fn ensure<'a, S>(
        &'a self, 
        table: &'a str, 
        id: &'a str, 
        value: &'a S
    ) -> EnsureFuture<'a, Self::Error>
    where
        S: Serialize + Send + Sync,
    { ... }

}

The backend to be used with a Database.

Associated Types

type Error: Send + Sync + StdError

The Error type that the backend will report up.

Required methods

fn has_table<'a>(&'a self, table: &'a str) -> HasTableFuture<'a, Self::Error>

Check if a table exists in the Database.

fn create_table<'a>(
&'a self,
table: &'a str
) -> CreateTableFuture<'a, Self::Error>

Inserts or creates a table in the Database.

fn delete_table<'a>(
&'a self,
table: &'a str
) -> DeleteTableFuture<'a, Self::Error>

Deletes or drops a table from the Database.

fn get_keys<'a, I>(
&'a self,
table: &'a str
) -> GetKeysFuture<'a, I, Self::Error> where
I: FromIterator<String>,

Gets all the keys in the table.

fn get<'a, D>(
&'a self,
table: &'a str,
id: &'a str
) -> GetFuture<'a, D, Self::Error> where
D: for<'de> Deserialize<'de> + Send + Sync,

Gets a certain entry from a table.

fn has<'a>(&'a self, table: &'a str, id: &'a str) -> HasFuture<'a, Self::Error>

Checks if an entry exists in a table.

fn create<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> CreateFuture<'a, Self::Error> where
S: Serialize + Send + Sync,

Inserts a new entry into a table.

fn update<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> UpdateFuture<'a, Self::Error> where
S: Serialize + Send + Sync,

Updates an existing entry in a table.

fn replace<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> ReplaceFuture<'a, Self::Error> where
S: Serialize + Send + Sync,

Replaces an existing entry in a table.

fn delete<'a>(
&'a self,
table: &'a str,
id: &'a str
) -> DeleteFuture<'a, Self::Error>

Deletes an entry from a table.

Provided methods

fn init(&self) -> InitFuture<'_, Self::Error>

An optional initialization function, useful for making connections to the database.

The default impl does nothing

unsafe fn shutdown(&self) -> ShutdownFuture<'_>

An optional shutdown function, useful for disconnecting from databases gracefully.

The default impl does nothing

Safety

This should not fail, as it’s ran upon dropping the Gateway, and panicking during a drop means resources haven’t adequately been cleaned up, which isn’t inherintly UB however it should still be documented.

fn ensure_table<'a>(
&'a self,
table: &'a str
) -> EnsureTableFuture<'a, Self::Error>

Ensures a table exists in the Database. Uses has_table first, then create_table if it returns false.

fn get_all<'a, D, I>(
&'a self,
table: &'a str,
entries: &'a [&str]
) -> GetAllFuture<'a, I, Self::Error> where
D: for<'de> Deserialize<'de> + Send + Sync,
I: FromIterator<D>,

Gets all entries that match a predicate, to get all entries, use get_keys first.

fn ensure<'a, S>(
&'a self,
table: &'a str,
id: &'a str,
value: &'a S
) -> EnsureFuture<'a, Self::Error> where
S: Serialize + Send + Sync,

Ensures a value exists in the table.