Struct starchart::Action

pub struct Action<S, C: CrudOperation, T: OpTarget> { /* fields omitted */ }

An Action for easy CRUD operations within a Starchart.

Implementations

impl<S: Entry, C: CrudOperation, T: OpTarget> Action<S, C, T>

pub fn new() -> Self

Creates a new Action with the specified operation.

pub fn kind(&self) -> ActionKind

Returns the ActionKind we will be performing with said action.

pub fn target(&self) -> OperationTarget

Returns the OperationTarget we will be performing with said action.

pub fn into_operation<O: CrudOperation>(self) -> Action<S, O, T>

Changes the CrudOperation of this Action.

pub fn into_target<T2: OpTarget>(self) -> Action<S, C, T2>

Changes the OpTarget of this Action.

pub fn into_create(self) -> Action<S, CreateOperation, T>

Sets the CrudOperation of this Action to CreateOperation.

pub fn into_read(self) -> Action<S, ReadOperation, T>

Sets the CrudOperation of this Action to ReadOperation.

pub fn into_update(self) -> Action<S, UpdateOperation, T>

Sets the CrudOperation of this Action to UpdateOperation.

pub fn into_delete(self) -> Action<S, DeleteOperation, T>

Sets the CrudOperation of this Action to DeleteOperation.

pub fn into_table(self) -> Action<S, C, TableTarget>

Sets the OpTarget of this Action to TableTarget.

pub fn into_entry(self) -> Action<S, C, EntryTarget>

Sets the OpTarget of this Action to EntryTarget.

pub fn with_entry<S2>(self) -> Action<S2, C, T>

Sets the target Entry of this Action.

pub fn set_table(&mut self, table_name: &str) -> &mut Self

Sets the table for this action.

impl<S: Entry, C: CrudOperation> Action<S, C, EntryTarget>

pub fn set_key<K: Key>(&mut self, key: &K) -> &mut Self

Sets the key for the action.

Users should prefer to call Self::set_entry over this, as setting the entry will automatically call this.

This is unused on OperationTarget::Table actions.

pub fn set_data(&mut self, entity: &S) -> &mut Self

Sets the data for the action.

This is unused on OperationTarget::Table actions.

impl<S: Entry, T: OpTarget> Action<S, CreateOperation, T>

pub fn create() -> Self

Begins a CreateOperation action.

impl<S: Entry, T: OpTarget> Action<S, ReadOperation, T>

pub fn read() -> Self

Begins a ReadOperation action.

impl<S: Entry, T: OpTarget> Action<S, UpdateOperation, T>

pub fn update() -> Self

Begins an UpdateOperation action.

impl<S: Entry, T: OpTarget> Action<S, DeleteOperation, T>

pub fn delete() -> Self

Begins a DeleteOperation action.

impl<S: Entry, C: CrudOperation> Action<S, C, TableTarget>

pub fn table() -> Self

Creates a new TableTarget based operation.

impl<S: Entry, C: CrudOperation> Action<S, C, EntryTarget>

pub fn entry() -> Self

Creates a new EntryTarget based operation.

impl<S: Entry> Action<S, CreateOperation, TableTarget>

pub fn create_table() -> Self

Creates a new CreateOperation based TableTarget operation.

impl<S: Entry> Action<S, ReadOperation, TableTarget>

pub fn read_table() -> Self

Creates a new ReadOperation based TableTarget operation.

impl<S: Entry> Action<S, DeleteOperation, TableTarget>

pub fn delete_table() -> Self

Creates a new DeleteOperation based TableTarget operation.

impl<S: Entry> Action<S, CreateOperation, EntryTarget>

pub fn create_entry() -> Self

Creates a new CreateOperation based EntryTarget operation.

impl<S: Entry> Action<S, ReadOperation, EntryTarget>

pub fn read_entry() -> Self

Creates a new ReadOperation based EntryTarget operation.

impl<S: Entry> Action<S, UpdateOperation, EntryTarget>

pub fn update_entry() -> Self

Creates a new UpdateOperation based EntryTarget operation.

impl<S: Entry> Action<S, DeleteOperation, EntryTarget>

pub fn delete_entry() -> Self

Creates a new DeleteOperation based EntryTarget operation.

impl<S: IndexEntry, C: CrudOperation> Action<S, C, EntryTarget>

pub fn set_entry(&mut self, entity: &S) -> &mut Self

Sets the Entry and Key that this Action will act over.